Fix async, major rewrite of most back ends. Major changes to tables and scripts.

+++ SQL Scripts/UpdateTicket.sql
@@ -1,4 +1,4 @@
-create definer = markus@`%` procedure UpdateTicket(IN TicketId int, IN WorkerId int, IN ThisTitle varchar(45), IN ThisDescription varchar(300), IN ThisResolution varchar(300), IN StatusInd tinyint(1))
+create definer = markus@`%` procedure UpdateTicket(IN TicketId int, IN WorkerId int, IN ThisTitle varchar(45), IN ThisDescription varchar(300), IN ThisResolution varchar(300), IN StatusInd int)
 begin
     update Ticket
     set worker_id   = WorkerId,
